+++ src/extension.ts
@@ -1,4 +1,4 @@
-import { commands, ExtensionContext, window, workspace } from "vscode";
+import { commands, ExtensionContext, window, workspace, env, Uri } from "vscode";
 import * as CaseConversion from "./modules/caseConversion";
 import * as InsertText from "./modules/insertText";
 import * as StatusBarSelection from "./modules/statusBarSelection";
@@ -12,6 +12,8 @@ import * as AlignText from "./modules/alignText";
 import TTDecorations from "./modules/decorations";
 import { tabOut, toggleTabOut } from "./modules/tabOut";
 import * as Delimiters from "./modules/delimiters";
+import * as path from "path";
+import { REGEX_VALIDATE_EMAIL } from "./modules/filterText";
 
 export function activate(context: ExtensionContext) {
     console.log("vscode-texttoolbox is active");
@@ -154,6 +156,54 @@ export function activate(context: ExtensionContext) {
     context.subscriptions.push(
         commands.registerTextEditorCommand("vscode-texttoolbox.OpenSelectionInNewEditor", () => {
             FilterText.openSelectionInNewEditor();
+        })
+    );
+
+    context.subscriptions.push(
+        commands.registerTextEditorCommand("vscode-texttoolbox.OpenPathOrUrlUnderCursor", async () => {
+            const editor = Helpers.getActiveEditor();
+            if (!editor) {
+                return;
+            }
+
+            let textBetweenSpaces = FilterText.getTextBetweenSpaces(editor);
+            if (textBetweenSpaces) {
+                // URL
+                if (textBetweenSpaces.startsWith("http") || textBetweenSpaces.startsWith("www.")) {
+                    if (!textBetweenSpaces.startsWith("http")) {
+                        textBetweenSpaces = "https://" + textBetweenSpaces;
+                    }
+                    env.openExternal(Uri.parse(textBetweenSpaces));
+                    return;
+                }
+
+                // Email
+                if (textBetweenSpaces.match(REGEX_VALIDATE_EMAIL)) {
+                    if (!textBetweenSpaces.startsWith("http")) {
+                        textBetweenSpaces = "mailto://" + textBetweenSpaces;
+                    }
+                    env.openExternal(Uri.parse(textBetweenSpaces));
+                    return;
+                }
+
+                // File system path
+                let userPathUri = Uri.file(textBetweenSpaces);
+                await workspace.fs.stat(userPathUri).then(
+                    (stat) => {
+                        commands.executeCommand("vscode.open", userPathUri);
+                    },
+                    (err) => {
+                        // if the path does not exist, check if it is a path relative to the open document
+                        let folder = path.dirname(editor.document.uri.fsPath);
+                        if (folder) {
+                            userPathUri = Uri.joinPath(Uri.file(folder), textBetweenSpaces!);
+                            workspace.fs.stat(userPathUri).then((stat) => {
+                                commands.executeCommand("vscode.open", userPathUri);
+                            });
+                        }
+                    }
+                );
+            }
         })
     );
 
@@ -406,12 +456,12 @@ export function activate(context: ExtensionContext) {
 
     // tabOut
     context.subscriptions.push(
-        commands.registerTextEditorCommand("vscode-texttoolbox.TabOut", () => { 
+        commands.registerTextEditorCommand("vscode-texttoolbox.TabOut", () => {
             tabOut();
         })
     );
     context.subscriptions.push(
-        commands.registerTextEditorCommand("vscode-texttoolbox.ToggleTabOut", () => { 
+        commands.registerTextEditorCommand("vscode-texttoolbox.ToggleTabOut", () => {
             toggleTabOut();
         })
     );